[Android] IncognitoStateProvider moved to chrome/browser/tabmodel
Chromium change:
https://chromium.googlesource.com/chromium/src/+/c9569b318d4fdde05de54bc5fe9e388c17554a42

Move IncognitoStateProvider to chrome/browser/tabmodel

IncognitoStateProvider uses a tab model to provide its incognito state,
used by toolbar/omnibox. This CL moves it to tabmodel module, in
an effort to reduce the omnibox's dependency on toolbar.

Bug: 1127732
